qai_hub_models/models/yolonas_quantized/requirements.txt
object-detection-metrics==0.4.post1
stringcase==1.2.0
rapidfuzz==3.8.1
treelib==1.6.1
imagesize==1.4.1
einops==0.3.2
Deprecated==1.2.11
data-gradients==0.3.1
shapely==2.0.3
boto3>=1.34,<1.36
torchmetrics==1.4.0.post0
onnxsim<=0.4.36
termcolor<=2.5.0
psutil>6,<7
jsonschema>4,<5
fiftyone==1.0.1
matplotlib==3.7.5 # Ultralytics does not specify a matplotlib verion, and the latest matplotlib breaks it. So we have to include the requirement manually
ultralytics==8.0.193 # used by shared YOLO app
